+++ R/pmcmc.R
@@ -4,7 +4,7 @@
 ##' `filter` is run with a set of parameters to evaluate the
 ##' likelihood. A new set of parameters is proposed, and these
 ##' likelihoods are compared, jumping with probability equal to their
-##' ratio. This is repeated for `n_mcmc` proposals.
+##' ratio. This is repeated for `n_steps` proposals.
 ##'
 ##' While this function is called `pmcmc` and requires a particle
 ##' filter object, there's nothing special about it for particle
@@ -48,6 +48,7 @@ pmcmc <- function(pars, filter, initial = NULL, control = NULL) {
   assert_is(pars, c("pmcmc_parameters", "pmcmc_parameters_nested"))
   assert_is(filter, c("particle_filter", "particle_deterministic"))
   assert_is(control, "pmcmc_control")
+  pmcmc_check_control(control)
   initial <- pmcmc_check_initial(initial, pars, control$n_chains)
 
   if (control$n_workers == 1) {