+++ R/pmcmc_tools.R
@@ -7,16 +7,17 @@
 ##' @param object Results of running [pmcmc()]
 ##'
 ##' @param burnin Optional integer number of iterations to discard as
-##'   "burn-in". If given then samples `1:burnin` will be
-##'   excluded from your results. Remember that the first sample
-##'   represents the starting point of the chain. It is an error if
-##'   this is not a positive integer or is greater than or equal to
-##'   the number of samples (i.e., there must be at least one sample
-##'   remaining after discarding burnin).
+##'   "burn-in". If given then samples `1:burnin` will be excluded
+##'   from your results. It is an error if this is not a positive
+##'   integer or is greater than or equal to the number of samples
+##'   (i.e., there must be at least one sample remaining after
+##'   discarding burnin).
 ##'
 ##' @param thin Optional integer thinning factor. If given, then every
-##'   `thin`'th sample is retained (e.g., if `thin` is 10
-##'   then we keep samples 1, 11, 21, ...).
+##'   `thin`'th sample is retained (e.g., if `thin` is 10 then we keep
+##'   samples 1, 11, 21, ...).  Note that this can produce surprising
+##'   results as it will always select the first sample but not
+##'   necessarily always the last.
 ##'
 ##' @export
 pmcmc_thin <- function(object, burnin = NULL, thin = NULL) {
@@ -30,7 +31,7 @@ pmcmc_thin <- function(object, burnin = NULL, thin = NULL) {
       stop(sprintf("'burnin' must be less than %d for your results",
                    burnin_max))
     }
-    i <- i & object$iteration >= burnin
+    i <- i & object$iteration > burnin
   }
 
   if (!is.null(thin)) {
@@ -81,6 +82,9 @@ pmcmc_filter <- function(object, i) {
     object$restart$state <- array_nth_dimension(object$restart$state, k, i)
   }
 
+  ## This must be removed (if it was present before)
+  object["pars_index"] <- list(NULL)
+
   object
 }
 
@@ -100,8 +104,8 @@ pmcmc_filter <- function(object, i) {
 pmcmc_combine <- function(..., samples = list(...)) {
   assert_list_of(samples, "mcstate_pmcmc")
 
-  pars <- lapply(samples, "[[", "pars")
-  probabilities <- lapply(samples, "[[", "probabilities")
+  pars <- lapply(samples, "[[", "pars_full")
+  probabilities <- lapply(samples, "[[", "probabilities_full")
   iteration <- lapply(samples, "[[", "iteration")
   state <- lapply(samples, "[[", "state")
   trajectories <- lapply(samples, "[[", "trajectories")
@@ -139,8 +143,8 @@ pmcmc_combine <- function(..., samples = list(...)) {
   ## We might check index, rate and step here though.
   predict <- last(samples)$predict
 
-  mcstate_pmcmc(pars, probabilities, state, trajectories, restart,
-                predict, chain, iteration)
+  mcstate_pmcmc(iteration, pars, probabilities, state, trajectories,
+                restart, predict, chain)
 }
 
 check_combine <- function(samples, iteration, state, trajectories, restart) {