+++ tests/testthat/test-pmcmc-utils.R
@@ -4,7 +4,7 @@ test_that("format and print the simplest object", {
   pars <- matrix(NA_real_, 10, 4,
                  dimnames = list(NULL, c("a", "b", "c", "d")))
   probs <- matrix(NA_real_, 10, 3, dimnames = list(NULL, c("x", "y", "z")))
-  x <- mcstate_pmcmc(pars, probs, NULL, NULL, NULL, NULL)
+  x <- mcstate_pmcmc(1:10, pars, probs, NULL, NULL, NULL, NULL)
 
   expected <- c(
     "<mcstate_pmcmc> (10 samples)",
@@ -30,7 +30,7 @@ test_that("format and print with state", {
   predict <- NULL
   restart <- list(date = 1, state = array(NA_real_, c(4, 10, 1)))
 
-  x <- mcstate_pmcmc(pars, probs, state, trajectories, restart, predict)
+  x <- mcstate_pmcmc(1:10, pars, probs, state, trajectories, restart, predict)
 
   expected <- c(
     "<mcstate_pmcmc> (10 samples)",
@@ -57,7 +57,7 @@ test_that("format and print nested object", {
   predict <- NULL
   restart <- list(date = 1, state = array(NA_real_, c(4, 2, 10, 1)))
 
-  x <- mcstate_pmcmc(pars, probs, state, trajectories, restart, predict)
+  x <- mcstate_pmcmc(1:10, pars, probs, state, trajectories, restart, predict)
 
   expected <- c(
     "<mcstate_pmcmc> (10 samples)",
@@ -80,14 +80,14 @@ test_that("print multichain object", {
   x <- pmcmc_combine(samples = example_sir_pmcmc2()$results)
 
   expected <- c(
-    "<mcstate_pmcmc> (93 samples across 3 chains)",
-    "  pars: 93 x 2 matrix of parameters",
+    "<mcstate_pmcmc> (90 samples across 3 chains)",
+    "  pars: 90 x 2 matrix of parameters",
     "    beta, gamma",
-    "  probabilities: 93 x 3 matrix of log-probabilities",
+    "  probabilities: 90 x 3 matrix of log-probabilities",
     "    log_prior, log_likelihood, log_posterior",
-    "  state: 5 x 93 matrix of final states",
-    "  trajectories: 3 x 93 x 101 array of particle trajectories",
-    "  restart: 5 x 93 x 1 array of particle restart state")
+    "  state: 5 x 90 matrix of final states",
+    "  trajectories: 3 x 90 x 101 array of particle trajectories",
+    "  restart: 5 x 90 x 1 array of particle restart state")
 
   expect_equal(format(x), expected)
   expect_output(print(x), paste(expected, collapse = "\n"), fixed = TRUE)
@@ -100,7 +100,7 @@ test_that("wrap long variable names nicely", {
   probs <- matrix(NA_real_, 10, 3, dimnames = list(NULL, c("x", "y", "z")))
   x <- withr::with_options(
     list(width = 80),
-    format(mcstate_pmcmc(pars, probs, NULL, NULL, NULL, NULL)))
+    format(mcstate_pmcmc(1:10, pars, probs, NULL, NULL, NULL, NULL)))
   expect_equal(
     x[[3]],
     "    aaaaaaaaaa, bbbbbbbbbbbbbbbbbbbb, cccccccccccccccccccccccccccccc,")